Fix NotEnoughBytes errors
The problem is that fill_buf doesn't always fill the buffer despite
what the name suggests. It only does so when the buffer (to be
consumed) is empty.

The buf_redux crate addresses this issue.

We use the MinBuffered policy as a temporay workaround but we should
instead do more precise buffer management. We should be able to
estimate minimum required bytes from the section length in the SHB
or block length in each block.